+++ save-cloud-common/src/commonMain/kotlin/com/saveourtool/save/cvsscalculator/VectorTypes.kt
@@ -1,6 +1,6 @@
 @file:Suppress("HEADER_MISSING_IN_NON_SINGLE_CLASS_FILE")
 
-package com.saveourtool.save.cosv.cvsscalculator
+package com.saveourtool.save.cvsscalculator
 
 import kotlinx.serialization.Serializable
